(feat) Added provider banner  extensionslot to page header and fixed config for renaming clinic to Home (#32)
