+++ packages/esm-imaging-orders-app/src/header/imagining-header.scss
@@ -1,12 +1,20 @@
+@use '@carbon/layout';
+@use '@carbon/type';
 @use '@carbon/colors';
 
-.pageHeader {
-  border-bottom: 1px solid colors.$gray-20;
-}
 
-.pageHeaderActions {
-  height: auto;
+.pageHeader {
+  @include type.type-style('body-compact-02');
+  height: layout.$spacing-12;
+  display: flex;
+  justify-content: space-between;
+  padding: layout.$spacing-05;
+  background: colors.$white-0;
+  border-bottom: 1px solid colors.$gray-20;  margin-bottom: layout.$spacing-05;
+  margin-bottom: layout.$spacing-05;
+}
+.providerBannerInfoSlot {
   display: flex;
   align-items: center;
-  justify-content: flex-end;
+  gap: layout.$spacing-05;
 }
