+++ packages/esm-procedure-orders-app/src/header/procedure-header.component.tsx
@@ -1,32 +1,20 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { Calendar, Location } from '@carbon/react/icons';
-import { useSession, formatDate } from '@openmrs/esm-framework';
+import { PageHeader, PageHeaderContent, ExtensionSlot } from '@openmrs/esm-framework';
 import ProcedureIllustration from './procedure-illustration.component';
 import styles from './procedure-header.scss';
 
 export const ProcedureHeader: React.FC = () => {
   const { t } = useTranslation();
-  const userSession = useSession();
-  const userLocation = userSession?.sessionLocation?.display;
 
   return (
-    <div className={styles.header}>
-      <div className={styles['left-justified-items']}>
-        <ProcedureIllustration />
-        <div className={styles['page-labels']}>
-          <p className={styles['page-name']}>{t('procedure', 'Procedures')}</p>
-        </div>
+    <>
+      <div className={`omrs-main-content`}>
+        <PageHeader data-testid="patient-queue-header">
+          <PageHeaderContent className={styles.pageHeader} illustration={<ProcedureIllustration />} title={t('procedure', 'Procedures')} />
+          <ExtensionSlot className={styles.providerBannerInfoSlot} name="provider-banner-info-slot" />
+        </PageHeader>
       </div>
-      <div className={styles['right-justified-items']}>
-        <div className={styles['date-and-location']}>
-          <Location size={16} />
-          <span className={styles.value}>{userLocation}</span>
-          <span className={styles.middot}>&middot;</span>
-          <Calendar size={16} />
-          <span className={styles.value}>{formatDate(new Date(), { mode: 'standard' })}</span>
-        </div>
-      </div>
-    </div>
+    </>
   );
 };
