+++ packages/esm-procedure-orders-app/src/header/procedure-header.scss
@@ -1,60 +1,22 @@
-@use '@carbon/styles/scss/spacing';
-@use '@carbon/styles/scss/type';
-@use '@openmrs/esm-styleguide/src/vars' as *;
+@use '@carbon/layout';
+@use '@carbon/type';
+@use '@carbon/colors';
 
-.header {
+
+.pageHeader {
   @include type.type-style('body-compact-02');
-  color: $text-02;
-  height: spacing.$spacing-12;
-  background-color: $ui-02;
-  border: 1px solid $ui-03;
-  border-left: 0px;
+  height: layout.$spacing-12;
   display: flex;
   justify-content: space-between;
+  padding: layout.$spacing-05;
+  background: colors.$white-0;
+  border-bottom: 1px solid colors.$gray-20;  margin-bottom: layout.$spacing-05;
+  margin-bottom: layout.$spacing-05;
 }
-
-.left-justified-items {
+.providerBannerInfoSlot {
   display: flex;
-  flex-direction: row;
   align-items: center;
-  margin-left: 0.75rem;
-}
-
-.right-justified-items {
-  @include type.type-style('body-compact-02');
-  color: $text-02;
-  padding-top: 1rem;
-}
-
-.page-name {
-  @include type.type-style('heading-04');
-}
-
-.page-labels {
-  margin-left: 1rem;
-
-  p:first-of-type {
-    margin-bottom: 0.25rem;
-  }
-}
-
-.date-and-location {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  margin-right: 1rem;
-}
-
-.value {
-  margin-left: 0.25rem;
-}
-
-.middot {
-  margin: 0 0.5rem;
-}
-
-.view {
-  @include type.type-style('label-01');
+  gap: layout.$spacing-05;
 }
 
 svg.iconOverrides {
@@ -67,4 +29,5 @@ svg.iconOverrides {
   width: 72px;
   height: 72px;
   fill: var(--brand-03);
+  margin-right: 1rem;
 }
